+++ qoqo-quest/src/backend.rs
@@ -15,14 +15,14 @@ use pyo3::exceptions::{PyRuntimeError, PyTypeError, PyValueError};
 use pyo3::ffi::c_str;
 use pyo3::prelude::*;
 use pyo3::types::{PyByteArray, PyType};
-use qoqo::convert_into_circuit;
-use qoqo::QoqoBackendError;
-use roqoqo::backends::EvaluatingBackend;
-use roqoqo::operations::*;
-use roqoqo::registers::{BitOutputRegister, ComplexOutputRegister, FloatOutputRegister};
-use roqoqo::Circuit;
+use qoqo::{convert_into_circuit, noise_models::ImperfectReadoutModelWrapper, QoqoBackendError};
+use roqoqo::{
+    backends::EvaluatingBackend,
+    operations::*,
+    registers::{BitOutputRegister, ComplexOutputRegister, FloatOutputRegister},
+    Circuit,
+};
 use std::collections::HashMap;
-
 /// QuEST backend
 ///
 /// provides functions to run circuits and measurements on with the QuEST quantum simulator.
@@ -195,6 +195,55 @@ impl BackendWrapper {
         warn_pragma_getstatevec_getdensitymat(circuit.clone());
         EvaluatingBackend::run_circuit(&self.internal, &circuit)
             .map_err(|err| PyRuntimeError::new_err(format!("Running Circuit failed {:?}", err)))
+    }
+
+    /// Run a circuit with the QuEST backend.
+    ///
+    /// A circuit is passed to the backend and executed.
+    /// During execution values are written to and read from classical registers
+    /// (List[bool], List[float], List[complex]).
+    /// To produce sufficient statistics for evaluating expectation values,
+    /// circuits have to be run multiple times.
+    /// The results of each repetition are concatenated in OutputRegisters
+    /// (List[List[bool]], List[List[float]], List[List[complex]]).
+    /// As a simulater Backend the QuEST backend also allows to direclty read out
+    /// the statevector, density matrix or the expectation values of products of PauliOperators
+    ///
+    ///
+    /// Args:
+    ///     circuit (Circuit): The circuit that is run on the backend.
+    ///     noise_model (NoiseModel): The noise model to be applied to the circuit.
+    /// Returns:
+    ///     Tuple[Dict[str, List[List[bool]]], Dict[str, List[List[float]]]], Dict[str, List[List[complex]]]]: The output registers written by the evaluated circuits.
+    ///
+    /// Raises:
+    ///     TypeError: Circuit argument cannot be converted to qoqo Circuit
+    ///     RuntimeError: Running Circuit failed
+    pub fn run_circuit_imperfect(
+        &self,
+        circuit: &Bound<PyAny>,
+        noise_model: &Bound<PyAny>,
+    ) -> PyResult<Registers> {
+        let circuit = convert_into_circuit(circuit).map_err(|err| {
+            PyTypeError::new_err(format!(
+                "Circuit argument cannot be converted to qoqo Circuit {:?}",
+                err
+            ))
+        })?;
+        let noise_model = match ImperfectReadoutModelWrapper::from_pyany(noise_model)? {
+            roqoqo::noise_models::NoiseModel::ImperfectReadoutModel(model) => model,
+            _ => return Err(PyTypeError::new_err("args: noise_model must be an instance of ImperfectReadoutModel, got {noise_model} instead."))
+        };
+        warn_pragma_getstatevec_getdensitymat(circuit.clone());
+        let (bit_register, float_register, complex_register) =
+            EvaluatingBackend::run_circuit(&self.internal, &circuit).map_err(|err| {
+                PyRuntimeError::new_err(format!("Running Circuit failed {:?}", err))
+            })?;
+        println!("AHAHAHAH");
+
+        let bit_register =
+            roqoqo_quest::apply_noisy_readouts(bit_register, &noise_model, self.get_random_seed());
+        Ok((bit_register, float_register, complex_register))
     }
 
     /// Run all circuits corresponding to one measurement with the QuEST backend.