+++ roqoqo-quest/src/interface/mod.rs
@@ -23,6 +23,8 @@ mod pragma_operations;
 use pragma_operations::*;
 mod gate_operations;
 use gate_operations::*;
+mod postprocessing;
+pub use postprocessing::apply_noisy_readouts;
 mod preprocessing;
 pub(crate) use pragma_operations::{
     execute_pragma_repeated_measurement, execute_replaced_repeated_measurement,