+++ roqoqo-quest/src/lib.rs
@@ -24,7 +24,9 @@
 //! roqoqo-quest provides a backend to simulate roqoqo quantum circuits with the QuEST simulator
 
 mod interface;
-pub use interface::{call_circuit, call_operation, call_operation_with_device};
+pub use interface::{
+    apply_noisy_readouts, call_circuit, call_operation, call_operation_with_device,
+};
 mod backend;
 pub use backend::Backend;
 mod quest_bindings;